+++ src/ensemble/random_forest/parameters.rs
@@ -1,8 +1,7 @@
 //! Public random-forest parameter types.
 
 use super::super::forest::forest_params;
-use crate::ensemble::MaxFeatures;
-use crate::tree::Splitter;
+use crate::tree::{MaxFeatures, Splitter};
 
 forest_params!(
     RandomForestClassifierParams,